[FIX]categ: do not assign a nul category to object without category (is_null(false) is false in php)
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@22624 b456876b-0849-0410-b77d-98878d47e9d5

+++ lib/categories/categlib.php
@@ -342,6 +342,9 @@ class CategLib extends ObjectLib {
 	}
 
 	function categorize($catObjectId, $categId) {
+		if (empty($categId)) {
+			return;
+		}
 		$query = "delete from `tiki_category_objects` where `catObjectId`=? and `categId`=?";
 		$result = $this->query($query,array((int) $catObjectId,(int) $categId),-1,-1,false);
 	        
@@ -1518,7 +1521,7 @@ class CategLib extends ObjectLib {
 		
 		if (empty($categories)) {
 			$forcedcat = $userlib->get_user_group_default_category($user);
-			if ( !is_null($forcedcat) ) {
+			if ( !empty($forcedcat) ) {
 				$categories[] = $forcedcat;
 			}
 		}
